+++ LookupAnything/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Added buildings to search UI.
+
 ## 1.45.3
 Released 07 July 2024 for SMAPI 4.0.7 or later.